implement support for --sigma_file input muchdogesec/siemrules#28; add date and modified #55
